+++ src/containers/home/ScheduleDrawer/ScheduleDrawer.tsx
@@ -1,23 +1,6 @@
-import {
-  Alert,
-  Box,
-  Button,
-  Slide,
-  SlideProps,
-  Snackbar,
-  Stack,
-} from "@mui/material";
+import { Alert, Box, Slide, SlideProps, Snackbar, Stack } from "@mui/material";
 import { useEffect, useRef, useState } from "react";
-import { useSelector } from "react-redux";
-import NameInput from "./inputs/NameInput";
-import DateInput from "./inputs/DateInput";
-import { SCHEDULE_DRAWER_MODE } from "../../../constants/schedule";
-import {
-  selectSchedule,
-  setDrawerSchedule,
-} from "../../../app/redux/slices/scheduleSlice";
-import CategoryInput from "./inputs/CategoryInput";
-import RepeatInput from "./inputs/RepeatInput";
+import { setDrawerSchedule } from "../../../app/redux/slices/scheduleSlice";
 import AssetFormPage from "./pages/AssetFormPage";
 import ScheduleDrawerHeader from "./layouts/ScheduleDrawerHeader";
 import ScheduleDrawerFooter from "./layouts/ScheduleDrawerFooter";
@@ -26,7 +9,7 @@ import { Schedule, ScheduleDrawerModeValue } from "../../../types/schedule";
 import { useAppDispatch } from "../../../app/redux/hooks";
 import { Swiper, SwiperSlide } from "swiper/react";
 import { Swiper as SwiperType } from "swiper/types";
-import ThickDivider from "@components/common/ThickDivider";
+import ScheduleFormPage from "@containers/home/ScheduleDrawer/pages/ScheduleFormPage/ScheduleFormPage.tsx";
 
 function TransitionUp(props: SlideProps) {
   return <Slide {...props} direction="right" />;
@@ -49,13 +32,11 @@ function ScheduleDrawer({
   const random = Math.floor(Math.random() * 4); // 현재 CONSUMPTION_ALERTS의 길이가 4임
 
   const dispatch = useAppDispatch();
-  const schedule = useSelector(selectSchedule);
 
   const [snackbarOpen, setSnackbarOpen] = useState(true);
   const [showError, setShowError] = useState(false);
   const [value, setValue] = useState(0);
   const [swiper, setSwiper] = useState<SwiperType>();
-  const [isCategoryPickerOpen, setIsCategoryPickerOpen] = useState(false);
 
   const handleChange = (event: React.SyntheticEvent, newValue: number) => {
     swiper?.slideTo(newValue);
@@ -83,87 +64,62 @@ function ScheduleDrawer({
 
   return (
     <div ref={ref} style={{ height: "100%" }}>
-      {schedule && (
-        <Box sx={{ height: "100%", pt: 1, mb: 3 }}>
-          <Snackbar
-            anchorOrigin={{ vertical: "top", horizontal: "center" }}
-            autoHideDuration={5000}
-            open={snackbarOpen}
-            onClose={() => {
-              setSnackbarOpen(false);
+      <Box sx={{ height: "100%", pt: 1, mb: 3 }}>
+        <Snackbar
+          anchorOrigin={{ vertical: "top", horizontal: "center" }}
+          autoHideDuration={5000}
+          open={snackbarOpen}
+          onClose={() => {
+            setSnackbarOpen(false);
+          }}
+          TransitionComponent={TransitionUp}
+        >
+          <Alert
+            color={CONSUMPTION_ALERTS[random].color}
+            sx={{ width: "100%" }}
+            icon={CONSUMPTION_ALERTS[random].icon}
+          >
+            {CONSUMPTION_ALERTS[random].message}
+          </Alert>
+        </Snackbar>
+
+        <ScheduleDrawerHeader
+          value={value}
+          handleChange={handleChange}
+          handleReset={handleReset}
+        />
+
+        <Stack
+          justifyContent="space-between"
+          spacing={2}
+          sx={{ height: `calc(100% - 56px)` }}
+          onTouchStart={(e) => e.stopPropagation()}
+        >
+          <Swiper
+            className="mySwiper"
+            style={{
+              width: "100%",
             }}
-            TransitionComponent={TransitionUp}
+            spaceBetween={50}
+            onSlideChange={(e) => setValue(e.activeIndex)}
+            onSwiper={(swiper) => setSwiper(swiper)}
           >
-            <Alert
-              color={CONSUMPTION_ALERTS[random].color}
-              sx={{ width: "100%" }}
-              icon={CONSUMPTION_ALERTS[random].icon}
-            >
-              {CONSUMPTION_ALERTS[random].message}
-            </Alert>
-          </Snackbar>
+            <SwiperSlide style={{ overflow: "scroll" }}>
+              <ScheduleFormPage mode={mode} showError={showError} />
+            </SwiperSlide>
+            <SwiperSlide>
+              <AssetFormPage mode={mode} />
+            </SwiperSlide>
+          </Swiper>
 
-          <ScheduleDrawerHeader
-            value={value}
-            handleChange={handleChange}
-            handleReset={handleReset}
+          {/* 제출 버튼 */}
+          <ScheduleDrawerFooter
+            mode={mode}
+            handleClose={handleClose}
+            setShowError={setShowError}
           />
-
-          <Stack
-            justifyContent="space-between"
-            spacing={2}
-            sx={{ height: `calc(100% - 56px)` }}
-            onTouchStart={(e) => e.stopPropagation()}
-          >
-            <Swiper
-              className="mySwiper"
-              style={{
-                width: "100%",
-              }}
-              spaceBetween={50}
-              onSlideChange={(e) => setValue(e.activeIndex)}
-              onSwiper={(swiper) => setSwiper(swiper)}
-            >
-              <SwiperSlide style={{ overflow: "scroll" }}>
-                <Stack spacing={2} pt={2}>
-                  {/* 이벤트 제목 */}
-                  <NameInput showError={showError} />
-
-                  {/* 이벤트 카테고리 */}
-                  <CategoryInput
-                    selected={
-                      mode === SCHEDULE_DRAWER_MODE.create
-                        ? ""
-                        : schedule.category
-                    }
-                    showError={showError}
-                  />
-                  <Button>일정 카테고리 선택하기 (임시)</Button>
-                  <ThickDivider />
-
-                  {/* 이벤트 일정 */}
-                  <DateInput showError={showError} />
-
-                  <ThickDivider />
-
-                  {/* 이벤트 반복 설정 */}
-                  <RepeatInput />
-                </Stack>
-              </SwiperSlide>
-              <SwiperSlide>
-                <AssetFormPage mode={mode} />
-              </SwiperSlide>
-            </Swiper>
-
-            {/* 제출 버튼 */}
-            <ScheduleDrawerFooter
-              mode={mode}
-              handleClose={handleClose}
-              setShowError={setShowError}
-            />
-          </Stack>
-        </Box>
-      )}
+        </Stack>
+      </Box>
     </div>
   );
 }
