+++ src/containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/CategoryPicker/CategoryPicker.tsx
