+++ src/containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/CategoryPicker/components/CategorySideTabButton.stories.tsx
