+++ src/containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/CategoryPicker/constants.ts
