+++ src/containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/CategoryPicker/index.ts
