+++ src/containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/CategoryPicker/pages/ExpenditureCategoryPage/index.ts
