+++ src/containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/CategoryPicker/pages/IncomeCategoryPage/IncomeCategoryPage.tsx
