+++ src/containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/CategoryPicker/pages/IncomeCategoryPage/index.ts
